// src/app/interfaces/grants.ts
export interface Grants {
  id?: number;
  membersread?: number;
  memberswrite?: number;
  membersadd?: number;
  membersupdate?: number;
  membersproductsadd?: number;
  memberspaymentscheduleread?: number;
  membersstatsread?: number;
  memberssubscriptionread?: number;
  paymentschedulesread?: number;
  paymentscheduleswrite?: number;
  paymentdayread?: number;
  drinksell?: number;
  foodsell?: number;
  sendnewsletter?: number;
}